Done code refactoring.
