+++ args.go
@@ -18,15 +18,19 @@ func (args CommandLineArgs) addOptionArgs(name string, value string) {
 	if args.optionArgs[name] == nil {
 		args.optionArgs[name] = make([]string, 0)
 	}
+
 	args.optionArgs[name] = append(args.optionArgs[name], value)
 }
 
 func (args CommandLineArgs) getOptionNames() []string {
 	argMapKeys := goo.GetType(args.optionArgs).GetGoValue().MapKeys()
+
 	mapKeys := make([]string, len(argMapKeys))
+
 	for i := 0; i < len(argMapKeys); i++ {
 		mapKeys[i] = argMapKeys[i].String()
 	}
+
 	return mapKeys
 }
